Update player-fixes-2 branch

+++ python/modules/Document.py
@@ -34,14 +34,17 @@ def RenderDocumentationFiles(aboutDir: str,destDir:str = "",pathToPrototype:str 
     sourceDir = aboutDir + "Sources"
     
     fileContents = {}
+    fileModified = {}
     for fileName in sorted(os.listdir(sourceDir)):
         sourcePath = Utils.PosixJoin(sourceDir,fileName)
 
         if not os.path.isfile(sourcePath) or not fileName.endswith(".md"):
             continue
 
+        fileModified[fileName] = Utils.ModificationDate(sourcePath)
         with open(sourcePath,encoding='utf8') as file:
-            fileContents[fileName] = file.read()
+            template = pyratemp.Template(file.read())
+            fileContents[fileName] = template(gOptions = gOptions,gDatabase = gDatabase)
             gDocumentationWordCount += WordCount(fileContents[fileName])
             
     def ApplyToText(transform: Callable[[str],Tuple[str,int]]) -> int:
@@ -59,7 +62,7 @@ def RenderDocumentationFiles(aboutDir: str,destDir:str = "",pathToPrototype:str 
     if html:
         htmlFiles = {}
         for fileName in fileContents:
-            html = markdown.markdown(fileContents[fileName],extensions = ["sane_lists","footnotes",NewTabRemoteExtension()])
+            html = markdown.markdown(fileContents[fileName],extensions = ["sane_lists","footnotes","toc",NewTabRemoteExtension()])
         
             html = re.sub(r"<!--HTML(.*?)-->",r"\1",html) # Remove comments around HTML code
             htmlFiles[Utils.ReplaceExtension(fileName,".html")] = html
@@ -77,6 +80,7 @@ def RenderDocumentationFiles(aboutDir: str,destDir:str = "",pathToPrototype:str 
 
         page = Html.PageDesc(Html.PageInfo(title,Utils.PosixJoin(destDir,fileName),titleInPage))
         page.AppendContent(fileText)
+        page.sourceFile = Utils.PosixJoin(sourceDir,Utils.ReplaceExtension(fileName,".md"))
         renderedPages.append(page)
 
     return renderedPages
@@ -102,10 +106,22 @@ def PrintWordCount() -> None:
 def AddArguments(parser) -> None:
     "Add command-line arguments used by this module"
     parser.add_argument('--documentationDir',type=str,default='documentation',help='Read and write documentation files here; Default: ./documenation')
+    parser.add_argument('--info',type=str,action="append",default=[],help="Specify infomation about this build. Format key:value")
     
 
 def ParseArguments() -> None:
-    pass
+    class NameSpace:
+        pass
+    infoObject = NameSpace()
+    for item in gOptions.info:
+        split = item.split(":")
+        if len(split) > 1:
+            value = split[1]
+        else:
+            value = True
+        setattr(infoObject,split[0],value)
+    gOptions.info = infoObject
+    
 
 def Initialize() -> None:
     pass
@@ -117,11 +133,15 @@ gDocumentationWordCount = 0
 def main() -> None:
     global gDocumentationWordCount
     gDocumentationWordCount = 0
+    modifiedFiles = []
     for directory in ['about','misc']:
         os.makedirs(Utils.PosixJoin(gOptions.documentationDir,directory),exist_ok=True)
         for page in RenderDocumentationFiles(directory,pathToPrototype=Utils.PosixJoin("../../",gOptions.prototypeDir),pathToBase="../../",html=False):
-            with open(page.info.file,'w',encoding='utf-8') as file:
-                print(str(page),file=file)
+            if not os.path.isfile(page.info.file) or Utils.DependenciesModified(page.info.file,[page.sourceFile]):
+                with open(page.info.file,'w',encoding='utf-8') as file:
+                    print(str(page),file=file)
+                modifiedFiles.append(page.info.file)
     
     if Alert.verbosity >= 2:
         PrintWordCount()
+    Alert.info("Wrote",len(modifiedFiles),".md files:",modifiedFiles)
